requirements.txt
Django>=1.6.0
django-reversion>=1.8.4
djangorestframework>=2.4.0
